272 dont allow user to delete themself

+++ apps/backend/src/system/users/users.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from "@nestjs/common";
+import { Controller, Get, Post, Body, Patch, Param, Delete, Session, ForbiddenException } from "@nestjs/common";
 import { UsersService } from "./users.service";
 import { ApiOperation, ApiTags, ApiUnprocessableEntityResponse } from "@nestjs/swagger";
 import { Roles } from "./roles.decorator";
@@ -6,6 +6,7 @@ import { UserRole } from "./user.entity";
 import { ErrorResponseDto } from "../../common/dto/error.dto";
 import { UserCreationDto, UserMutationDto, UserResponseDto } from "./user.dto";
 import { Serialize } from "../../common/decorators/serialize";
+import { Session as FastifySession } from "@fastify/secure-session";
 
 @ApiTags("Users")
 @Controller("users")
@@ -44,7 +45,9 @@ export class UsersController {
   @Delete(":id")
   @ApiOperation({ summary: "Delete a user by ID" })
   @Roles(UserRole.ADMIN)
-  delete(@Param("id") id: string) {
+  delete(@Param("id") id: string, @Session() session: FastifySession) {
+    if (session.get("userId") === +id) throw new ForbiddenException("You mustn't delete yourself.");
+
     return this.usersService.remove(+id);
   }
 }
